Build Cypress tests with WebPack
