+++ package.json
@@ -50,6 +50,10 @@
   },
   "devDependencies": {
     "@4tw/cypress-drag-drop": "^2.2.1",
+    "@babel/core": "^7.18.9",
+    "@babel/preset-env": "^7.18.9",
+    "@cypress/webpack-batteries-included-preprocessor": "^2.2.3",
+    "@cypress/webpack-preprocessor": "^5.12.0",
     "@octokit/rest": "^19.0.1",
     "@snowpack/app-scripts-react": "^2.0.1",
     "@snowpack/plugin-postcss": "^1.4.3",
@@ -74,6 +78,7 @@
     "@typescript-eslint/eslint-plugin": "^5.30.7",
     "@typescript-eslint/parser": "^5.30.7",
     "@vitejs/plugin-react": "^1.3.2",
+    "babel-loader": "^8.2.5",
     "cypress": "^10.3.1",
     "decompress": "^4.2.1",
     "del": "^7.0.0",
@@ -98,7 +103,8 @@
     "ts-node": "^10.9.1",
     "typescript": "^4.7.4",
     "vite": "^2.9.14",
-    "vitest": "^0.19.0"
+    "vitest": "^0.19.0",
+    "webpack": "^5.73.0"
   },
   "lint-staged": {
     "*.{js,jsx,mjs,ts,tsx}": "eslint --cache --fix"
